refactor(babel-plugin-espower): it's better to create capturer in BabelAssertionVisitor
since BabelAssertionVisitor knows when and where to create it
