+++ lib/babel-assertion-visitor.js
@@ -1,11 +1,46 @@
 'use strict';
 
 var espower = require('espower');
+var estraverse = require('estraverse');
 var inherits = require('util').inherits;
 var babelgen = require('babel-generator');
 var define = require('define-properties');
 var toBeCaptured = require('./to-be-captured');
 var toBeSkipped = require('./to-be-skipped');
+var helperCode = [
+    '(function () {',
+    '    var captured = [];',
+    '    function _capt (value, espath) {',
+    '        captured.push({value: value, espath: espath});',
+    '        return value;',
+    '    }',
+    '    function _expr (value, args) {',
+    '        var source = {',
+    '            content: args.content,',
+    '            filepath: args.filepath,',
+    '            line: args.line',
+    '        };',
+    '        if (args.generator) {',
+    '            source.generator = true;',
+    '        }',
+    '        if (args.async) {',
+    '            source.async = true;',
+    '        }',
+    '        return {',
+    '            powerAssertContext: {',
+    '                value: value,',
+    '                events: captured',
+    '            },',
+    '            source: source',
+    '        };',
+    '    }',
+    '    return {',
+    '        _capt: _capt,',
+    '        _expr: _expr',
+    '    };',
+    '});'
+].join('\n');
+
 
 function BabelAssertionVisitor (babel, matcher, nodePath, createEsPath, enclosingFunc, options) {
     espower.AssertionVisitor.call(this, matcher, createEsPath(nodePath), enclosingFunc, options);
@@ -13,6 +48,8 @@ function BabelAssertionVisitor (babel, matcher, nodePath, createEsPath, enclosin
     this.createEsPath = createEsPath;
     this.assertionBabelNodePath = nodePath;
     this.currentArgumentBabelNodePath = null;
+    var babelTemplate = babel.template;
+    this.helperTemplate = babelTemplate(helperCode);
 }
 inherits(BabelAssertionVisitor, espower.AssertionVisitor);
 
@@ -80,6 +117,10 @@ BabelAssertionVisitor.prototype.enterArgument = function (babelNodePath) {
             // skip optional message argument
             return undefined;
         }
+
+        // create capturer per argument
+        this.powerAssertCalleeObject = this.createNewCapturer(babelNodePath);
+
         this.verifyNotInstrumented(currentNode);
         // entering target argument
         this.currentArgumentBabelNodePath = babelNodePath;
@@ -108,6 +149,46 @@ BabelAssertionVisitor.prototype.isLeavingAssertion = function (babelNodePath) {
 
 BabelAssertionVisitor.prototype.isLeavingArgument = function (babelNodePath) {
     return this.currentArgumentBabelNodePath === babelNodePath;
+};
+
+BabelAssertionVisitor.prototype.createNewCapturer = function (nodePath) {
+    var types = this.babel.types;
+    var helperNameNode = this.getCaptureHelperNameNode(nodePath);
+    var capturerIdent = nodePath.scope.generateUidIdentifier('rec');
+    define(capturerIdent, {_generatedByEspower: true});
+    var init = types.callExpression(helperNameNode, []);
+    define(init, {_generatedByEspower: true});
+    nodePath.scope.push({id: capturerIdent, init: init});
+    return capturerIdent;
+};
+
+BabelAssertionVisitor.prototype.getCaptureHelperNameNode = function (nodePath) {
+    var file = nodePath.hub.file;
+    var helperNameNode = file.get('powerAssertCaptureHelper');
+    if (!helperNameNode) {
+        helperNameNode = this.createHelperNameNode(nodePath);
+        // helperNameNode = file.addImport('power-assert-runtime/recorder', 'default', 'recorder');
+    }
+    return helperNameNode;
+};
+
+BabelAssertionVisitor.prototype.createHelperNameNode = function (nodePath) {
+    var types = this.babel.types;
+    var file = nodePath.hub.file;
+    var programScope = nodePath.scope.getProgramParent();
+    var helperNameNode = programScope.generateUidIdentifier('powerAssertRecorder');
+    define(helperNameNode, {_generatedByEspower: true});
+    file.set('powerAssertCaptureHelper', helperNameNode);
+    var helperFunctionNode = types.toExpression(this.helperTemplate());
+    estraverse.traverse(helperFunctionNode, {
+        keys: types.VISITOR_KEYS,
+        enter: function (currentNode) {
+            define(currentNode, {_generatedByEspower: true});
+        }
+    });
+    helperFunctionNode._compact = true;
+    programScope.push({id: helperNameNode, init: helperFunctionNode});
+    return helperNameNode;
 };
 
 module.exports = BabelAssertionVisitor;
