+++ lib/babel-espower-visitor.js
@@ -7,48 +7,12 @@ var extend = require('xtend');
 var define = require('define-properties');
 var find = require('array-find');
 var BabelAssertionVisitor = require('./babel-assertion-visitor');
-var helperCode = [
-    '(function () {',
-    '    var captured = [];',
-    '    function _capt (value, espath) {',
-    '        captured.push({value: value, espath: espath});',
-    '        return value;',
-    '    }',
-    '    function _expr (value, args) {',
-    '        var source = {',
-    '            content: args.content,',
-    '            filepath: args.filepath,',
-    '            line: args.line',
-    '        };',
-    '        if (args.generator) {',
-    '            source.generator = true;',
-    '        }',
-    '        if (args.async) {',
-    '            source.async = true;',
-    '        }',
-    '        return {',
-    '            powerAssertContext: {',
-    '                value: value,',
-    '                events: captured',
-    '            },',
-    '            source: source',
-    '        };',
-    '    }',
-    '    return {',
-    '        _capt: _capt,',
-    '        _expr: _expr',
-    '    };',
-    '});'
-].join('\n');
-
 
 function BabelEspowerVisitor (babel, opts) {
     this.babel = babel;
     this.esTreePath = require('./babel-estree-path')(babel);
     this.matchers = opts.patterns.map(function (pattern) { return escallmatch(pattern, opts); });
     this.options = opts;
-    var babelTemplate = babel.template;
-    this.helperTemplate = babelTemplate(helperCode);
 }
 
 BabelEspowerVisitor.prototype.enter = function (nodePath) {
@@ -68,9 +32,6 @@ BabelEspowerVisitor.prototype.enter = function (nodePath) {
         if (!assertionVisitor.isCapturingArgument() && !this.isCalleeOfParentCallExpression(parentNode, currentKey)) {
             // entering argument
             assertionVisitor.enterArgument(nodePath);
-            if (assertionVisitor.isCapturingArgument()) {
-                assertionVisitor.powerAssertCalleeObject = this.createNewCapturer(nodePath);
-            }
         }
     } else if (types.isCallExpression(currentNode)) {
         var matcher = find(this.matchers, function (m) { return m.test(currentNode); });
@@ -138,45 +99,6 @@ BabelEspowerVisitor.prototype.exit = function (nodePath) {
     if (isTreeModified) {
         nodePath.replaceWith(resultTree);
     }
-};
-
-BabelEspowerVisitor.prototype.createNewCapturer = function (nodePath) {
-    var types = this.babel.types;
-    var helperNameNode = this.getCaptureHelperNameNode(nodePath);
-    var capturerIdent = nodePath.scope.generateUidIdentifier('rec');
-    define(capturerIdent, {_generatedByEspower: true});
-    var init = types.callExpression(helperNameNode, []);
-    define(init, {_generatedByEspower: true});
-    nodePath.scope.push({id: capturerIdent, init: init});
-    return capturerIdent;
-};
-
-BabelEspowerVisitor.prototype.getCaptureHelperNameNode = function (nodePath) {
-    var file = nodePath.hub.file;
-    var helperNameNode = file.get('powerAssertCaptureHelper');
-    if (!helperNameNode) {
-        helperNameNode = this.createHelperNameNode(nodePath);
-    }
-    return helperNameNode;
-};
-
-BabelEspowerVisitor.prototype.createHelperNameNode = function (nodePath) {
-    var types = this.babel.types;
-    var file = nodePath.hub.file;
-    var programScope = nodePath.scope.getProgramParent();
-    var helperNameNode = programScope.generateUidIdentifier('powerAssertRecorder');
-    define(helperNameNode, {_generatedByEspower: true});
-    file.set('powerAssertCaptureHelper', helperNameNode);
-    var helperFunctionNode = types.toExpression(this.helperTemplate());
-    estraverse.traverse(helperFunctionNode, {
-        keys: types.VISITOR_KEYS,
-        enter: function (currentNode) {
-            define(currentNode, {_generatedByEspower: true});
-        }
-    });
-    helperFunctionNode._compact = true;
-    programScope.push({id: helperNameNode, init: helperFunctionNode});
-    return helperNameNode;
 };
 
 BabelEspowerVisitor.prototype.isCalleeOfParentCallExpression = function (parentNode, currentKey) {
